requirements.txt
ipython
jupyter
notebook
numpy~=1.21.4
pandas~=1.3.4
seaborn
pytest
tqdm
numpy
pandas
matplotlib
pathlib2
torch~=1.10.0
fastai
pytest-warnings
Pillow~=8.4.0